+++ src/app/page.tsx
@@ -1,10 +1,3 @@
-import CommentList from '@/components/common/CommentList';
-
 export default function Home() {
-  return (
-    <>
-      <div>랜딩페이지</div>
-      <CommentList />
-    </>
-  );
+  return <div>랜딩페이지</div>;
 }
